ADD: abstract super class for npc & player properties
