+++ Assets/Scripts/NPC.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 29b37aa485fa4ad9a23523dcfe13f377
+timeCreated: 1589021641
