+++ Assets/Scripts/NPC/CharacterProperties.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: fb737bd803cc4fe496b73f60455acff5
+timeCreated: 1589021755
