+++ Assets/Scripts/NPC/NPCProperties.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace NPC
+{
+    public class NpcProperties : CharacterProperties
+    {
+        private void Start()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
