+++ Assets/Scripts/NPC/NPCProperties.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 209c8f3ff3934c418ad8995b27e49987
+timeCreated: 1589021642
